+++ integrations/tendermint-grpc/tsconfig.json
@@ -7,5 +7,6 @@
     "esModuleInterop": true,
     "resolveJsonModule": true
   },
-  "files": ["src/index.ts"]
+  "files": ["src/index.ts"],
+  "assets": ["src/runtime.proto"]
 }
